--- library/src/wu/seal/jsontokotlin/test/JsonToKotlinBuilderTest.kt
@@ -1,11 +1,0 @@
-package wu.seal.jsontokotlin.test
-
-import org.junit.jupiter.api.Test
-
-class JsonToKotlinBuilderTest {
-
-    @Test
-    fun test() {
-
-    }
-}
